drivers: cir: Expand durations to 16-bit
Currently all durations fit in signed 8-bit variable. However, this will
change in following commit with higher clock rate.

Make durations 16-bit.

+++ drivers/cir/rc6.c
@@ -33,7 +33,7 @@ enum {
 	RC6_STATES
 };
 
-static const int8_t rc6_durations[RC6_STATES] = {
+static const int16_t rc6_durations[RC6_STATES] = {
 	[RC6_IDLE]      = RC6_UNITS_TO_CLKS(6),
 	[RC6_LEADER_S]  = RC6_UNITS_TO_CLKS(2),
 	[RC6_HEADER_P]  = RC6_UNITS_TO_CLKS(1),
